Bug fixes and issues resolved
Resolved the issue #10 
Steps taken to resolve that issue are:
 - Setting a maximum limit of salary a job can have
 - Formatting numbers by appending a suffix (ie: "K", "B", "M")

Bug fixes: Got the"attempting to index a nil value" error when trying to choose a job that was created by a company owner and that job is the second one created by him. The root for the case was that company doesn't updated in the tempData table for the relevant player. Solved by keeping that value without removing after a job creation!